Weaviate/requirements.txt
transformers
sentence-transformers
torch
tqdm
weaviate-client